cleaned up code and set up api to work in hybrid mode running locally and connecting to google cloud sql
