+++ src/database.py
@@ -6,13 +6,10 @@ from sqlalchemy import (
     ForeignKey,
     Date,
     Boolean,
-    create_engine,
 )
-from sqlalchemy.orm import declarative_base, sessionmaker
+from sqlalchemy.orm import declarative_base
 from sqlalchemy.dialects.postgresql import JSON
 
-# engine = create_engine("postgresql://katcha@localhost:5432/erg_track", echo=False)
-# Session = sessionmaker(bind=engine)
 
 Base = declarative_base()
 
@@ -60,19 +57,3 @@ class WorkoutLogTable(Base):
 
     def __repr__(self):
         return f"<WorkoutLogTable(workout_id={self.workout_id}, user_id={self.user_id}, date={self.date}, time={self.time}, meter={self.meter}, split={self.split}, stroke_rate={self.stroke_rate}, interval={self.interval}, image_hash={self.image_hash}, subworkouts={self.subworkouts}, comment={self.comment})>"
-
-
-# Below:  uses SQLAlchemy directly to make tables, confuses aleembic. Don't use/
-# Base.metadata.create_all(engine)
-
-
-# class SubWorkoutTable(Base):
-#     __tablename__ = "subworkout"
-
-#     sub_id = Column(Integer, Sequence("subworkout_id_sequ"), primary_key=True)
-#     workout_id = Column(Integer, ForeignKey("workout_log.workout_id"))
-#     sub_time = Column(String)
-#     sub_meter = Column(Integer)
-#     sub_split = Column(String)
-#     sub_stroke_rate = Column(Integer)
-#     rest = Column(String)
